moved email.ts back to saltcorn-data and changed deps
 - removed the saltcorn-email package

 - changed the github dependency to the npm registry
