--- packages/saltcorn-email/package.json
@@ -1,66 +1,0 @@
-{
-  "name": "@saltcorn/email",
-  "version": "0.6.4-beta.2",
-  "description": "Email model for Saltcorn, open-source no-code platform",
-  "homepage": "https://saltcorn.com",
-  "scripts": {
-    "test": "jest tests --runInBand",
-    "tsc": "echo \"Error: no TypeScript support yet\"",
-    "clean": "echo \"Error: no TypeScript support yet\""
-  },
-  "author": "Tom Nielsen",
-  "license": "MIT",
-  "main": "dist/index.js",
-  "exports": {
-    "./*": "./dist/*.js"
-  },
-  "typesVersions": {
-    "*": {
-      "*": [
-        "dist/*"
-      ]
-    }
-  },
-  "files": [
-    "dist/**/*"
-  ],
-  "optionalDependencies": {
-    "@saltcorn/postgres": "0.6.4-beta.2",
-    "@saltcorn/sqlite": "0.6.4-beta.2"
-  },
-  "dependencies": {
-    "@saltcorn/data": "0.6.4-beta.2",
-    "@saltcorn/markup": "0.6.4-beta.2",
-    "bootstrap-email": "git+https://github.com/christianhugoch/bootstrap-email.git",
-    "nodemailer": "^6.4.17",
-    "tmp-promise": "^3.0.2",
-    "uuid": "^8.2.0"
-  },
-  "repository": "github:saltcorn/saltcorn",
-  "devDependencies": {
-    "@types/node": "^16.11.12",
-    "@types/nodemailer": "^6.4.4",
-    "@types/uuid": "^8.3.3",
-    "jest": "^27.1.0",
-    "ts-jest": "^27.1.0",
-    "typescript": "^4.4.4"
-  },
-  "jest": {
-    "preset": "ts-jest",
-    "testEnvironment": "node",
-    "modulePathIgnorePatterns": [
-      ".*\\.js$"
-    ],
-    "moduleNameMapper": {
-      "@saltcorn/sqlite/(.*)": "@saltcorn/sqlite/dist/$1",
-      "@saltcorn/db-common/(.*)": "@saltcorn/db-common/dist/$1",
-      "@saltcorn/data/(.*)": "@saltcorn/data/dist/$1",
-      "@saltcorn/types/(.*)": "@saltcorn/types/dist/$1",
-      "@saltcorn/markup$": "@saltcorn/markup/dist",
-      "@saltcorn/markup/(.*)": "@saltcorn/markup/dist/$1"
-    }
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
