+++ packages/saltcorn-random-tests/package.json
@@ -36,8 +36,7 @@
       "@saltcorn/types/(.*)": "@saltcorn/types/dist/$1",
       "@saltcorn/markup$": "@saltcorn/markup/dist",
       "@saltcorn/markup/(.*)": "@saltcorn/markup/dist/$1",
-      "@saltcorn/admin-models/(.*)": "@saltcorn/admin-models/dist/$1",
-      "@saltcorn/email/(.*)": "@saltcorn/email/dist/$1"
+      "@saltcorn/admin-models/(.*)": "@saltcorn/admin-models/dist/$1"
     }
   },
   "publishConfig": {
